+++ elements/drawtools/test/cases/load-drawtool.js
@@ -7,10 +7,8 @@ const { drawTools } = TEST_SELECTORS;
  * Test to verify if the drawtools component loads successfully.
  */
 const loadDrawToolsTest = () => {
-  it("loads the drawtools", () => {
-    // Find the drawTools element and access its shadow DOM
-    cy.get(drawTools).shadow();
-  });
+  // Find the drawTools element and access its shadow DOM
+  cy.get(drawTools).shadow();
 };
 
 export default loadDrawToolsTest;